pyproject.toml
[project]
name = "devex-slackbot"
version = "0.1.0"
description = "A Slack bot that answers questions using a Notion FAQ page"
requires-python = ">=3.10"
dependencies = [
    "slack-bolt>=1.18.0",
    "notion-client>=2.2.1",
    "anthropic>=0.18.0",
    "sentence-transformers>=2.3.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
